+++ fpsim/locations/ethiopia/regions/somali.py
@@ -15,9 +15,9 @@ thisdir = sc.thispath(__file__)  # For loading CSV files
 
 def scalar_pars():
     scalar_pars = eth.scalar_pars()
-    scalar_pars['location'] = 'amhara'
-    scalar_pars['breastfeeding_dur_mu'] = 12.5669
-    scalar_pars['breastfeeding_dur_beta'] = 10.75467
+    scalar_pars['location'] = 'somali'
+    scalar_pars['breastfeeding_dur_mu'] = 6.26613513894357
+    scalar_pars['breastfeeding_dur_beta'] = 5.53459861844008
 
     return scalar_pars
 
@@ -36,43 +36,43 @@ def filenames():
 
 # %% Demographics and pregnancy outcome
 '''
-Data from 1994 Census Report for Amhara Region
-https://www.statsethiopia.gov.et/wp-content/uploads/2019/06/Population-and-Housing-Census-1994-Amhara-Region.pdf
+Data from 1994 Census Report for Somali Region
+https://www.statsethiopia.gov.et/wp-content/uploads/2019/06/Population-and-Housing-Census-1994-Somali-Region.pdf
 '''
 def age_pyramid():
-    pyramid = np.array([[0, 1063425, 1039862], # Amhara 1994 
-                        [5, 1083627, 1072118],
-                        [10, 962649, 885335],
-                        [15, 742766, 740028],
-                        [20, 525138, 564480],
-                        [25, 481007, 531523],
-                        [30, 366492, 420412],
-                        [35, 342041, 367473],
-                        [40, 292577, 309360],
-                        [45, 248468, 211516],
-                        [50, 224383, 236333],
-                        [55, 149310, 120785],
-                        [60, 164406, 154539],
-                        [65, 109904, 78971],
-                        [70, 92528, 80203],
-                        [75, 45471, 31283],
-                        [80, 53354, 42530]
+    pyramid = np.array([[0, 236849, 209061], # Somali 1994 
+                        [5, 354039, 286088],
+                        [10, 365196, 267903],
+                        [15, 248098, 173908],
+                        [20, 146947, 114943],
+                        [25, 86685, 101700],
+                        [30, 80398, 101909],
+                        [35, 57492, 78121],
+                        [40, 80933, 78380],
+                        [45, 41953, 36890],
+                        [50, 53574, 40614],
+                        [55, 19403, 11803],
+                        [60, 37735, 18706],
+                        [65, 9732, 4490],
+                        [70, 15531, 7237],
+                        [75, 2951, 1430],
+                        [80, 7901, 4565]
                         ], dtype=float)    
     return pyramid
 
-def urban_proportion(): # TODO: Flagging this - currently this is being used for the urban ratio for amhara; I assume you'd want to use the region-specific value?
+def urban_proportion(): # TODO: Flagging this - currently this is being used for the urban ratio for somali; I assume you'd want to use the region-specific value?
     urban_data = eth.urban_proportion()
     
     return urban_data  # Return this value as a float
 
 def region_proportions():
     '''
-    Defines the proportion of the population in the Amhara region to establish the probability of living there.
+    Defines the proportion of the population in the Somali region to establish the probability of living there.
     '''
     region_data = pd.read_csv(thisdir / '..' / 'ethiopia' / 'data' / 'region.csv')
     region_dict = {}
-    region_dict['mean'] = region_data.loc[region_data['region'] == 'Amhara']['mean'] #HOLDER FOR NOW. NEED TO CALL ON LOCATION.
-    region_dict['urban'] = region_data.loc[region_data['region'] == 'Amhara']['urban']
+    region_dict['mean'] = region_data.loc[region_data['region'] == 'Somali']['mean'] #HOLDER FOR NOW. NEED TO CALL ON LOCATION.
+    region_dict['urban'] = region_data.loc[region_data['region'] == 'Somali']['urban']
     
     return region_dict
 
@@ -115,13 +115,13 @@ def fecundity_ratio_nullip():
 
 def lactational_amenorrhea_region():
     '''
-    Returns a dictionary containing the percent of breastfeeding women by month postpartum 0-11 months who meet criteria for LAM, specifically for the Amhara region.
+    Returns a dictionary containing the percent of breastfeeding women by month postpartum 0-11 months who meet criteria for LAM, specifically for the Somali region.
     '''
     lam_region = pd.read_csv(thisdir / '..' / 'ethiopia' / 'data' / 'lam_region.csv')
     lam_dict = {}
-    lam_dict['month'] = lam_region.loc[lam_region['region'] == 'Amhara']['month'].tolist()
+    lam_dict['month'] = lam_region.loc[lam_region['region'] == 'Somali']['month'].tolist()
     lam_dict['month'] = np.array(lam_dict['month'], dtype=np.float64)
-    lam_dict['rate'] = lam_region.loc[lam_region['region'] == 'Amhara']['rate'].tolist()
+    lam_dict['rate'] = lam_region.loc[lam_region['region'] == 'Somali']['rate'].tolist()
     lam_dict['rate'] = np.array(lam_dict['rate'], dtype=np.float64)
  
 
@@ -135,9 +135,9 @@ def sexual_activity_region(): #NEEDS UPDATING
     '''
     sexually_active_region_data = pd.read_csv(thisdir / '..' / 'ethiopia' / 'data' / 'sexual_activity_region.csv')
     sexually_active_region_dict = {}
-    sexually_active_region_dict['age'] = sexually_active_region_data.loc[sexually_active_region_data['region']== 'Amhara']['age'].tolist()   # Return age
+    sexually_active_region_dict['age'] = sexually_active_region_data.loc[sexually_active_region_data['region']== 'Somali']['age'].tolist()   # Return age
     sexually_active_region_dict['age'] = np.array(sexually_active_region_dict['age'], dtype=np.float64)
-    sexually_active_region_dict['perc'] = [x / 100 for x in sexually_active_region_data.loc[sexually_active_region_data['region']== 'Amhara']['perc'].tolist()]
+    sexually_active_region_dict['perc'] = [x / 100 for x in sexually_active_region_data.loc[sexually_active_region_data['region']== 'Somali']['perc'].tolist()]
     sexually_active_region_dict['perc'] = np.array(sexually_active_region_dict['perc'], dtype=np.float64)
 
     activity_interp_model_region = si.interp1d(x=sexually_active_region_dict['age'], y=sexually_active_region_dict['perc'])
@@ -151,9 +151,9 @@ def sexual_activity_pp_region():
     '''
     pp_activity_region = pd.read_csv(thisdir / '..' / 'ethiopia' / 'data' / 'sexual_activity_pp_region.csv')
     pp_activity_region_dict = {}
-    pp_activity_region_dict['month'] = pp_activity_region.loc[pp_activity_region['region'] == 'Amhara']['month'].tolist()
+    pp_activity_region_dict['month'] = pp_activity_region.loc[pp_activity_region['region'] == 'Somali']['month'].tolist()
     pp_activity_region_dict['month'] = np.array(pp_activity_region_dict['month'], dtype=np.float64)
-    pp_activity_region_dict['percent_active'] = pp_activity_region.loc[pp_activity_region['region'] == 'Amhara']['perc'].tolist()
+    pp_activity_region_dict['percent_active'] = pp_activity_region.loc[pp_activity_region['region'] == 'Somali']['perc'].tolist()
     pp_activity_region_dict['percent_active'] = np.array(pp_activity_region_dict['percent_active'], dtype=np.float64)
     
     return pp_activity_region_dict
@@ -164,9 +164,9 @@ def debut_age_region():
     '''
     sexual_debut_region_data = pd.read_csv(thisdir / '..' / 'ethiopia' / 'data' / 'sexual_debut_region.csv')
     debut_age_region_dict = {}
-    debut_age_region_dict['ages'] = sexual_debut_region_data.loc[sexual_debut_region_data['region'] == 'Amhara']['age'].tolist()
+    debut_age_region_dict['ages'] = sexual_debut_region_data.loc[sexual_debut_region_data['region'] == 'Somali']['age'].tolist()
     debut_age_region_dict['ages'] = np.array(debut_age_region_dict['ages'], dtype=np.float64)
-    debut_age_region_dict['probs'] = sexual_debut_region_data.loc[sexual_debut_region_data['region'] == 'Amhara']['prob'].tolist()
+    debut_age_region_dict['probs'] = sexual_debut_region_data.loc[sexual_debut_region_data['region'] == 'Somali']['prob'].tolist()
     debut_age_region_dict['probs'] = np.array(debut_age_region_dict['probs'], dtype=np.float64)
 
     return debut_age_region_dict
@@ -179,7 +179,7 @@ def exposure_age():
     also miscarriage), will decrease factor number
     '''
     exposure_correction_age = np.array([[0, 5, 10, 12.5, 15, 18, 20, 25, 30, 35, 40, 45, 50],
-                                        [1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1]])
+                                        [1, 1, 1, 1, 1, 1, 1, 0.8, 0.8, 0.8, 1, 1, 1]])
     exposure_age_interp = data2interp(exposure_correction_age, fpd.spline_preg_ages)
 
     return exposure_age_interp
@@ -398,8 +398,8 @@ def barriers_region():
 
     reasons_region = pd.read_csv(thisdir / '..' / 'ethiopia' / 'data' / 'barriers_region.csv')
     reasons_region_dict = {}
-    barriers = reasons_region.loc[reasons_region['region'] == 'Amhara']['barrier'].tolist() # Return the reason for nonuse
-    percs = reasons_region.loc[reasons_region['region'] == 'Amhara']['perc'].tolist() # Return the percentage
+    barriers = reasons_region.loc[reasons_region['region'] == 'Somali']['barrier'].tolist() # Return the reason for nonuse
+    percs = reasons_region.loc[reasons_region['region'] == 'Somali']['perc'].tolist() # Return the percentage
 
     for i in range(len(barriers)):
         reasons_region_dict[barriers[i]] = percs[i]