+++ CHANGELOG.rst
@@ -8,8 +8,14 @@ All notable changes to the codebase are documented in this file. Changes that ma
    :local:
    :depth: 1
 
+Version 0.28.1 (2024-04-11)
+---------------------------
+- Updates regional parameters to calibrate subnationally in Ethiopia
+- Revises subnational calibration script to run for all regions and includes a new multisim script for subnational comparisons 
+- *GitHub info*: PR `https://github.com/fpsim/fpsim/pull/319>`_
+
 Version 0.27.0 (2024-02-09)
---------------------------
+---------------------------
 - Builds out new regional attribute and parameters for subnational dynamics in Ethiopia
 - Creates new subnational calibration script and subfolder for regional datasets and figures 
 - *GitHub info*: PR `https://github.com/fpsim/fpsim/pull/241>`_